# backend/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Styling with Muskan API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/bookings": {
      "get": {
        "summary": "List my bookings",
        "parameters": [{ "name": "page", "in": "query", "schema": { "type": "integer" } }, { "name": "limit", "in": "query", "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "401": { "description": "Unauthorized" } }
      },
      "post": {
        "summary": "Create booking",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "items": { "type": "array", "items": { "type": "object" } },
                  "slot": { "type": "object", "properties": { "date": { "type": "string" }, "time": { "type": "string" } }, "required": ["date", "time"] },
                  "address": { "type": "object", "properties": { "houseNo": { "type": "string" }, "area": { "type": "string" }, "landmark": { "type": "string" }, "lat": { "type": "number" }, "lng": { "type": "number" } }, "required": ["houseNo", "area"] },
                  "bookingType": { "type": "string" },
                  "couponCode": { "type": "string" }
                },
                "required": ["items", "slot", "address", "bookingType"]
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Validation errors" }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/admin/settings": {
      "put": {
        "summary": "Update office settings",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "startTime": { "type": "string" }, "endTime": { "type": "string" }, "autoAssign": { "type": "boolean" } }, "required": ["startTime", "endTime", "autoAssign"] } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Validation errors" }, "401": { "description": "Unauthorized" } }
      }
    }
  }
}